requirements.txt
# Análisis de datos y manipulación
pandas>=2.0.0
numpy>=1.24.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Big Data y Spark
pyspark>=3.4.0
findspark>=2.0.1

# Conectores de bases de datos - MySQL optimizado
mysql-connector-python>=8.0.33
PyMySQL>=1.0.2
sqlalchemy>=2.0.0
cryptography>=41.0.0  # Para conexiones seguras MySQL

# Notebooks y desarrollo
jupyter>=1.0.0
ipykernel>=6.22.0

# Utilidades
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# API (opcional)
fastapi>=0.95.0
uvicorn>=0.21.0

# Dashboard (opcional)
streamlit>=1.22.0
dash>=2.10.0

# Base de datos (opcional)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
